% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnswerHelper.R
\docType{class}
\name{AnswerHelper}
\alias{AnswerHelper}
\title{Answer Helper Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnswerHelper
}
\description{
Answer Helper Class
}
\keyword{datasets}
