% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baseball.R
\docType{class}
\name{Baseball}
\alias{Baseball}
\title{Baseball Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
# Initialize a class
baseball <- Baseball$new()
# Start the game
baseball$start()
}
\description{
Number Baseball R6Class. \code{new()} and \code{start()} to begin the game
}
\section{Methods}{

\describe{
  \item{\code{new()}}{Constructor}
  \item{\code{start()}}{Start the game}
 }
}

\keyword{Game}
