% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInput.R
\docType{data}
\name{UserInput}
\alias{UserInput}
\title{UserInput Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserInput
}
\description{
UserInput Class
}
\keyword{datasets}
